source/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25
httpx>=0.26.0
beautifulsoup4>=4.12.3
lxml>=5.1.0
apscheduler>=3.10.4
pydantic>=2.5.3
pydantic-settings>=2.1.0
jinja2>=3.1.3
python-multipart>=0.0.6
aiosmtplib>=3.0.1
winotify>=1.1.0
pystray>=0.19.5
pillow>=10.2.0

pywebview>=4.4
